% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eirt_utilities.R
\name{coef.gllamm_eirt}
\alias{coef.gllamm_eirt}
\title{Extract Item Parameters from EIRT Model}
\usage{
\method{coef}{gllamm_eirt}(object, se = FALSE, ...)
}
\arguments{
\item{object}{A gllamm_eirt object}

\item{se}{Logical; include standard errors?}
}
\value{
A data frame with item parameters
}
\description{
Extract item parameters with standard errors (if available)
}
