% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eirt_utilities.R
\name{compare_eirt}
\alias{compare_eirt}
\title{Compare Explanatory IRT Models}
\usage{
compare_eirt(..., test = c("LRT", "none"))
}
\arguments{
\item{...}{Two or more gllamm_eirt objects to compare}

\item{test}{Type of test: "LRT" for likelihood ratio test, "none" for just IC comparison}
}
\value{
A data frame with model comparison statistics
}
\description{
Compare two or more EIRT models using likelihood ratio tests and information criteria
}
\details{
This function compares EIRT models to test the importance of item covariates.
Models are compared using:
\itemize{
  \item Log-likelihood
  \item AIC and BIC
  \item Likelihood ratio test (if nested models)
}

For nested models (e.g., with and without a predictor), the LRT tests
whether adding the predictor significantly improves fit.
}
\examples{
\dontrun{
# Fit model with predictor
fit1 <- fit_eirt(responses, item_data,
                 difficulty_formula = ~ word_freq,
                 model = "Rasch")

# Fit model without predictor
fit0 <- fit_eirt(responses, item_data,
                 difficulty_formula = ~ 1,
                 model = "Rasch")

# Compare models
compare_eirt(fit0, fit1)
}

}
