% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmb_interface.R
\name{compile_gllamm_tmb}
\alias{compile_gllamm_tmb}
\title{Compile TMB template}
\usage{
compile_gllamm_tmb(template = "gllamm_gaussian", dir = "src")
}
\arguments{
\item{template}{Name of template file (without .hpp extension)}

\item{dir}{Source directory (default: "src")}
}
\value{
TRUE if successful
}
\description{
Helper function to compile the C++ TMB template
}
