% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_multinomial.R
\name{compute_multinomial_probs}
\alias{compute_multinomial_probs}
\title{Compute multinomial probabilities}
\usage{
compute_multinomial_probs(X, beta, eta_random, n_categories)
}
\arguments{
\item{X}{Fixed effects design matrix}

\item{beta}{Beta matrix (K-1) × p}

\item{eta_random}{Random effects contribution (vector of length n_obs)}

\item{n_categories}{Number of categories}
}
\value{
Matrix of probabilities (n_obs × n_categories)
}
\description{
Compute multinomial probabilities
}
\keyword{internal}
