% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_statistics.R
\name{compute_test_information_summary}
\alias{compute_test_information_summary}
\title{Compute test information summary}
\usage{
compute_test_information_summary(object)
}
\description{
Compute test information summary
}
\keyword{internal}
