% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_random.R
\name{create_grouping_matrix}
\alias{create_grouping_matrix}
\title{Create grouping factor matrix}
\usage{
create_grouping_matrix(terms, data)
}
\arguments{
\item{terms}{Parsed random effects terms}

\item{data}{Person-level data}
}
\value{
List with group_ids matrix and group information
}
\description{
Create grouping factor matrix
}
\keyword{internal}
