% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_random.R
\name{extract_random_terms}
\alias{extract_random_terms}
\title{Extract random effects terms from formula}
\usage{
extract_random_terms(formula)
}
\description{
Extract random effects terms from formula
}
\keyword{internal}
