% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marginal_utils.R
\name{extract_random_vcov}
\alias{extract_random_vcov}
\title{Extract random effects variance-covariance matrix from fitted model}
\usage{
extract_random_vcov(object)
}
\arguments{
\item{object}{Fitted gllamm object}
}
\value{
Variance-covariance matrix of random effects
}
\description{
Extract random effects variance-covariance matrix from fitted model
}
\keyword{internal}
