% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{find_outliers}
\alias{find_outliers}
\title{Outlier detection for GLLAMM}
\usage{
find_outliers(object, threshold = 3, ...)
}
\arguments{
\item{object}{A gllamm object}

\item{threshold}{Threshold for standardized residuals (default: 3)}

\item{...}{Additional arguments}
}
\value{
Indices of potential outliers
}
\description{
Outlier detection for GLLAMM
}
