% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_statistics.R
\name{fit.gllamm}
\alias{fit.gllamm}
\title{Fit Statistics for Standard GLLAMM Models}
\usage{
\method{fit}{gllamm}(object, quiet = FALSE, ...)
}
\arguments{
\item{object}{A gllamm object}

\item{quiet}{Suppress ICC computation messages (default: FALSE)}

\item{...}{Additional arguments (not currently used)}
}
\value{
Object of class \code{fit_gllamm} with fit statistics
}
\description{
Fit Statistics for Standard GLLAMM Models
}
