% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_statistics.R
\name{fit.gllamm_irt}
\alias{fit.gllamm_irt}
\title{Fit Statistics for IRT Models}
\usage{
\method{fit}{gllamm_irt}(object, compute_item_fit = TRUE, compute_person_fit = TRUE, ...)
}
\arguments{
\item{object}{A gllamm_irt object}

\item{compute_item_fit}{Compute item fit statistics (S-X²) (default: TRUE)}

\item{compute_person_fit}{Compute person fit statistics (outfit/infit) (default: TRUE)}

\item{...}{Additional arguments}
}
\value{
Object of class \code{fit_irt} with IRT-specific fit statistics
}
\description{
Fit Statistics for IRT Models
}
