% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_statistics.R
\name{fit.gllamm_lca}
\alias{fit.gllamm_lca}
\title{Fit Statistics for Latent Class Analysis}
\usage{
\method{fit}{gllamm_lca}(object, ...)
}
\arguments{
\item{object}{A gllamm_lca object}

\item{...}{Additional arguments}
}
\value{
Object of class \code{fit_lca} with LCA-specific fit statistics
}
\description{
Fit Statistics for Latent Class Analysis
}
