% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_statistics.R
\name{fit.gllamm_multinomial}
\alias{fit.gllamm_multinomial}
\title{Fit Statistics for Multinomial Regression Models}
\usage{
\method{fit}{gllamm_multinomial}(object, ...)
}
\arguments{
\item{object}{A gllamm_multinomial object}

\item{...}{Additional arguments}
}
\value{
Object of class \code{fit_multinomial} with multinomial-specific fit statistics
}
\description{
Fit Statistics for Multinomial Regression Models
}
