% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_statistics.R
\name{fit.gllamm_ordinal}
\alias{fit.gllamm_ordinal}
\title{Fit Statistics for Ordinal Regression Models}
\usage{
\method{fit}{gllamm_ordinal}(object, test_po = TRUE, ...)
}
\arguments{
\item{object}{A gllamm_ordinal object}

\item{test_po}{Test proportional odds assumption (default: TRUE for logit/probit)}

\item{...}{Additional arguments}
}
\value{
Object of class \code{fit_ordinal} with ordinal-specific fit statistics
}
\description{
Fit Statistics for Ordinal Regression Models
}
