% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_statistics.R
\name{fit}
\alias{fit}
\title{Generic Fit Statistics Function}
\usage{
fit(object, ...)
}
\arguments{
\item{object}{A fitted model object}

\item{...}{Additional arguments passed to methods}
}
\value{
An object of class \code{fit_statistics} with model-specific components
}
\description{
Compute model-specific fit statistics for GLLAMM objects
}
\details{
The \code{fit()} function provides comprehensive, model-specific fit statistics:

\strong{GLMM models:}
\itemize{
  \item Log-likelihood, AIC, BIC
  \item R-squared (marginal and conditional for Gaussian models)
  \item Intraclass correlation coefficient (ICC)
}

\strong{IRT models:}
\itemize{
  \item Log-likelihood, AIC, BIC
  \item Item fit statistics (S-X^2)
  \item Person fit statistics (outfit/infit)
  \item Reliability estimates
  \item Test information function
}

\strong{Latent Class Analysis:}
\itemize{
  \item Log-likelihood, AIC, BIC
  \item Entropy (classification quality)
  \item Class proportions
  \item Average posterior probabilities (APPA)
}

\strong{Ordinal models:}
\itemize{
  \item Log-likelihood, AIC, BIC
  \item Pseudo-R^2 (McFadden)
  \item Proportional odds test (for PO/probit models)
}
}
\examples{
\dontrun{
# GLMM
fit1 <- gllamm(y ~ x + (1 | group), data = data)
fit(fit1)

# IRT
fit2 <- fit_irt(responses, model = "2PL")
fit(fit2, compute_item_fit = TRUE)

# LCA
fit3 <- fit_lca(indicators, nclass = 3)
fit(fit3)

# Ordinal
fit4 <- fit_ordinal(rating ~ x + (1 | id), data = data)
fit(fit4, test_po = TRUE)
}

}
