% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irt.R
\name{fit_irt_polytomous}
\alias{fit_irt_polytomous}
\title{Internal function for polytomous IRT models}
\usage{
fit_irt_polytomous(response_matrix, model, weights, re_info, start, control)
}
\description{
Internal function for polytomous IRT models
}
\keyword{internal}
