% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{fixef.gllamm}
\alias{fixef.gllamm}
\title{Extract fixed effects}
\usage{
\method{fixef}{gllamm}(object, ...)
}
\arguments{
\item{object}{A gllamm object}

\item{...}{Additional arguments (ignored)}
}
\value{
Named vector of fixed effects coefficients
}
\description{
Extract fixed effects
}
