% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marginal_utils.R
\name{get_inverse_link}
\alias{get_inverse_link}
\title{Get inverse link function for a family}
\usage{
get_inverse_link(family)
}
\arguments{
\item{family}{Family object (standard R family or custom)}
}
\value{
Inverse link function
}
\description{
Get inverse link function for a family
}
\keyword{internal}
