% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{gllamm-class}
\alias{gllamm-class}
\alias{print.gllamm}
\alias{summary.gllamm}
\alias{coef.gllamm}
\alias{vcov.gllamm}
\alias{logLik.gllamm}
\alias{fitted.gllamm}
\alias{residuals.gllamm}
\title{GLLAMM model object class}
\usage{
print.gllamm(x, ...)

summary.gllamm(object, ...)

\method{coef}{gllamm}(object, ...)

\method{vcov}{gllamm}(object, which = "fixed", ...)

\method{logLik}{gllamm}(object, ...)

\method{fitted}{gllamm}(object, ...)

\method{residuals}{gllamm}(object, type = c("response", "pearson", "deviance"), ...)
}
\arguments{
\item{x}{A gllamm object}

\item{...}{Additional arguments}

\item{object}{A gllamm object}
}
\description{
S3 class for fitted GLLAMM models
}
