% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{gof.gllamm}
\alias{gof.gllamm}
\title{Goodness-of-fit tests for GLLAMM}
\usage{
gof.gllamm(object, ...)
}
\arguments{
\item{object}{A gllamm object}

\item{...}{Additional arguments}
}
\description{
Goodness-of-fit tests for GLLAMM
}
