% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{icc.gllamm}
\alias{icc.gllamm}
\title{Variance decomposition (ICC)}
\usage{
\method{icc}{gllamm}(object, quiet = FALSE, ...)
}
\arguments{
\item{object}{A gllamm object}

\item{quiet}{Suppress printed output and messages (default: FALSE)}

\item{...}{Additional arguments}
}
\description{
Variance decomposition (ICC)
}
