% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{influence.gllamm}
\alias{influence.gllamm}
\title{Influence diagnostics for GLLAMM}
\usage{
\method{influence}{gllamm}(object, ...)
}
\arguments{
\item{object}{A gllamm object}

\item{...}{Additional arguments}
}
\value{
Data frame with influence measures
}
\description{
Influence diagnostics for GLLAMM
}
