% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marginal_utils.R
\name{mc_integrate_fixed_samples}
\alias{mc_integrate_fixed_samples}
\title{Monte Carlo integration for marginal predictions}
\usage{
mc_integrate_fixed_samples(X, Z, beta, u_samples, inv_link_fn)
}
\arguments{
\item{X}{Fixed effects design matrix (n x p)}

\item{Z}{Random effects design matrix (n x q) for a single observation}

\item{beta}{Fixed effects coefficients (p x 1)}

\item{u_samples}{Matrix of random effects samples (n_sim x q)}

\item{inv_link_fn}{Inverse link function}
}
\value{
Vector of marginal predictions (length n)
}
\description{
Computes E[g^(-1)(X'beta + Z'u)] by averaging over draws from u ~ N(0, Sigma_u)
}
\keyword{internal}
