% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marginal_utils.R
\name{mc_integrate_marginal}
\alias{mc_integrate_marginal}
\title{Monte Carlo integration for marginal predictions (one sample at a time)}
\usage{
mc_integrate_marginal(X, Z, beta, Sigma_u, inv_link_fn, n_sim = 1000)
}
\arguments{
\item{X}{Fixed effects design matrix}

\item{Z}{Random effects design matrix}

\item{beta}{Fixed effects coefficients}

\item{Sigma_u}{Random effects variance-covariance matrix}

\item{inv_link_fn}{Inverse link function}

\item{n_sim}{Number of Monte Carlo samples}
}
\value{
List with fit and se
}
\description{
More memory-efficient version that processes samples sequentially
}
\keyword{internal}
