% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aghq.R
\name{npml}
\alias{npml}
\title{Nonparametric maximum likelihood integration}
\usage{
npml(k = 2)
}
\arguments{
\item{k}{Number of mass points (default 2)}
}
\value{
An object of class \code{gllamm_integration}
}
\description{
Integration specification for \code{\link{gllamm}}: replace the normal
latent distribution with k estimated mass points and masses (NPML;
Aitkin 1999). Equivalent to \code{\link{fit_npml}}.
}
\examples{
\dontrun{
gllamm(y ~ x + (1 | g), data = d, family = binomial(),
       integration = npml(2))
}
}
