% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_random.R
\name{parse_single_random_term}
\alias{parse_single_random_term}
\title{Parse a single random effects term}
\usage{
parse_single_random_term(term, data)
}
\description{
Parse a single random effects term
}
\keyword{internal}
