% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ordinal.R
\name{plot_category_probs_ordinal}
\alias{plot_category_probs_ordinal}
\title{Plot Category Probabilities}
\usage{
plot_category_probs_ordinal(x, covariate, covariate_values, ...)
}
\description{
Plot Category Probabilities
}
\keyword{internal}
