% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ordinal.R
\name{plot_covariate_effects_ordinal}
\alias{plot_covariate_effects_ordinal}
\title{Plot Covariate Effects}
\usage{
plot_covariate_effects_ordinal(x, covariate, ...)
}
\description{
Plot Covariate Effects
}
\keyword{internal}
