% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ordinal.R
\name{plot_cumulative_probs_ordinal}
\alias{plot_cumulative_probs_ordinal}
\title{Plot Cumulative Probabilities}
\usage{
plot_cumulative_probs_ordinal(x, covariate, covariate_values, ...)
}
\description{
Plot Cumulative Probabilities
}
\keyword{internal}
