% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eirt_utilities.R
\name{plot_item_covariates}
\alias{plot_item_covariates}
\title{Plot Item Covariate Effects}
\usage{
plot_item_covariates(
  object,
  covariate,
  parameter = c("difficulty", "discrimination"),
  ...
)
}
\arguments{
\item{object}{A gllamm_eirt object}

\item{covariate}{Name of covariate to plot}

\item{parameter}{Which parameter to plot: "difficulty" or "discrimination"}

\item{...}{Additional arguments passed to plot}
}
\description{
Visualize the relationship between item covariates and item parameters
}
\examples{
\dontrun{
fit <- fit_eirt(responses, item_data,
                difficulty_formula = ~ word_freq)

plot_item_covariates(fit, covariate = "word_freq")
}

}
