% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict_marginal_gllamm}
\alias{predict_marginal_gllamm}
\title{Internal function for marginal predictions}
\usage{
predict_marginal_gllamm(object, newdata = NULL, n_sim = 1000, se.fit = FALSE)
}
\arguments{
\item{object}{Fitted gllamm object}

\item{newdata}{New data for predictions (NULL = use original data)}

\item{n_sim}{Number of Monte Carlo samples}

\item{se.fit}{Return standard errors?}
}
\value{
Vector of marginal predictions, or list with fit and se.fit
}
\description{
Computes population-averaged predictions by integrating over random effects
}
\keyword{internal}
