% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_ordinal.R
\name{predict_marginal_ordinal}
\alias{predict_marginal_ordinal}
\title{Internal function for marginal ordinal predictions}
\usage{
predict_marginal_ordinal(object, X, Z, n_sim = 1000)
}
\arguments{
\item{object}{Fitted ordinal model}

\item{X}{Fixed effects design matrix}

\item{Z}{Random effects design matrix}

\item{n_sim}{Number of MC samples}
}
\value{
Matrix of marginal probabilities (n_obs × n_categories)
}
\description{
Internal function for marginal ordinal predictions
}
\keyword{internal}
