% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/families.R
\name{print.binomial_family}
\alias{print.binomial_family}
\title{Print method for binomial family}
\usage{
\method{print}{binomial_family}(x, ...)
}
\description{
Print method for binomial family
}
\keyword{internal}
