% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dif.R
\name{print.dif_analysis}
\alias{print.dif_analysis}
\title{Print DIF analysis results}
\usage{
print.dif_analysis(x, ...)
}
\arguments{
\item{x}{Object of class dif_analysis}

\item{...}{Additional arguments (not used)}
}
\description{
Print DIF analysis results
}
