% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eirt.R
\name{print.gllamm_eirt}
\alias{print.gllamm_eirt}
\title{Print EIRT model results}
\usage{
print.gllamm_eirt(x, ...)
}
\arguments{
\item{x}{Object of class gllamm_eirt}

\item{...}{Additional arguments}
}
\description{
Print EIRT model results
}
