% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dif.R
\name{summary.dif_analysis}
\alias{summary.dif_analysis}
\title{Summary of DIF analysis}
\usage{
summary.dif_analysis(object, ...)
}
\arguments{
\item{object}{Object of class dif_analysis}

\item{...}{Additional arguments (not used)}
}
\description{
Summary of DIF analysis
}
