% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eirt.R
\name{summary.gllamm_eirt}
\alias{summary.gllamm_eirt}
\title{Summary of EIRT model}
\usage{
\method{summary}{gllamm_eirt}(object, ...)
}
\arguments{
\item{object}{Object of class gllamm_eirt}

\item{...}{Additional arguments}
}
\description{
Summary of EIRT model
}
