% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{VarCorr.gllamm}
\alias{VarCorr.gllamm}
\title{Extract variance components}
\usage{
\method{VarCorr}{gllamm}(object, ...)
}
\arguments{
\item{object}{A gllamm object}

\item{...}{Additional arguments (ignored)}
}
\value{
List of variance-covariance matrices for random effects
}
\description{
Extract variance components
}
