% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R, R/multilevel_methods.R
\name{VarCorr}
\alias{VarCorr}
\title{Generic VarCorr}
\usage{
VarCorr(x, ...)

VarCorr(x, ...)
}
\arguments{
\item{x}{A fitted model object}

\item{...}{Additional arguments passed to methods}

\item{object}{A fitted model object}
}
\value{
Variance components (method-specific format)
}
\description{
Generic function to extract variance components from fitted multi-level models
}
